Move DI related logic around clients and url to environment package
